requirements.txt
streamlit==1.22.0
pdfplumber==0.9.0
python-docx==0.8.11
python-dotenv==1.0.0
cohere==4.22
pandas==2.0.1